+++ frontend/src/app/core/services/events.service.ts
@@ -1,9 +1,9 @@
 import {Injectable} from '@angular/core';
-import {EventDto, EventsApiService} from "../../../core/api";
-import {ApplicationStateService} from "../../../core/services/application-state.service";
+import {EventDto, EventsApiService} from "../api";
+import {ApplicationStateService} from "./application-state.service";
 import {EventInput} from "@fullcalendar/core";
 import {EventMapperService} from "./event-mapper.service";
-import {SnackbarService} from "../../../core/services/snackbar.service";
+import {SnackbarService} from "./snackbar.service";
 
 @Injectable()
 export class EventsService {
